+++ src/autocomplete_test.js
@@ -13,6 +13,17 @@ var Range = require("./range").Range;
 require("./ext/language_tools");
 var Autocomplete = require("./autocomplete").Autocomplete;
 
+var MouseEvent = function(type, opts){
+    var e = document.createEvent("MouseEvents");
+    e.initMouseEvent(/click|wheel/.test(type) ? type : "mouse" + type,
+        true, true, window,
+        opts.detail,
+        opts.x, opts.y, opts.x, opts.y,
+        opts.ctrl, opts.alt, opts.shift, opts.meta,
+        opts.button || 0, opts.relatedTarget);
+    return e;
+};
+
 var editor;
 function initEditor(value) {
     if (editor) {
@@ -32,6 +43,13 @@ function initEditor(value) {
     return editor;
 }
 
+function afterRenderCheck(popup, callback) {
+    popup.renderer.on("afterRender", function wait() {
+        popup.renderer.off("afterRender", wait);
+        callback();
+    });
+}
+
 module.exports = {
     tearDown: function() {
         if (editor) {
@@ -48,16 +66,16 @@ module.exports = {
         assert.ok(!editor.container.querySelector("style"));
 
         sendKey("a");
-        checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-label="arraysort" aria-setsize="2" aria-posinset="0" aria-describedby="doc-tooltip"><s "ace_completion-highlight">a</s><s "ace_">rraysort</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d><d "ace_line"><s "ace_completion-highlight">a</s><s "ace_">looooooooooooooooooooooooooooong_word</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
+        checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-label="arraysort" aria-setsize="2" aria-posinset="1" aria-describedby="doc-tooltip"><s "ace_completion-highlight">a</s><s "ace_">rraysort</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d><d "ace_line"><s "ace_completion-highlight">a</s><s "ace_">looooooooooooooooooooooooooooong_word</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
             sendKey("rr");
-            checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-label="arraysort" aria-setsize="1" aria-posinset="0" aria-describedby="doc-tooltip"><s "ace_completion-highlight">arr</s><s "ace_">aysort</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
+            checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-label="arraysort" aria-setsize="1" aria-posinset="1" aria-describedby="doc-tooltip"><s "ace_completion-highlight">arr</s><s "ace_">aysort</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
                 sendKey("r");
-                checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-label="arraysort" aria-setsize="1" aria-posinset="0" aria-describedby="doc-tooltip"><s "ace_completion-highlight">arr</s><s "ace_">ayso</s><s "ace_completion-highlight">r</s><s "ace_">t</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
+                checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-label="arraysort" aria-setsize="1" aria-posinset="1" aria-describedby="doc-tooltip"><s "ace_completion-highlight">arr</s><s "ace_">ayso</s><s "ace_completion-highlight">r</s><s "ace_">t</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
                     
                     sendKey("Return");
                     assert.equal(editor.getValue(), "arraysort\narraysort alooooooooooooooooooooooooooooong_word");
                     editor.execCommand("insertstring", " looooooooooooooooooooooooooooong_");
-                    checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-label="alooooooooooooooooooooooooooooong_word" aria-setsize="1" aria-posinset="0" aria-describedby="doc-tooltip"><s "ace_">a</s><s "ace_completion-highlight">looooooooooooooooooooooooooooong_</s><s "ace_">word</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
+                    checkInnerHTML('<d "ace_line ace_selected" id="suggest-aria-id:0" role="option" aria-label="alooooooooooooooooooooooooooooong_word" aria-setsize="1" aria-posinset="1" aria-describedby="doc-tooltip"><s "ace_">a</s><s "ace_completion-highlight">looooooooooooooooooooooooooooong_</s><s "ace_">word</s><s "ace_completion-spacer"> </s><s "ace_completion-meta">local</s></d>', function() {
                         sendKey("Return");
                         editor.destroy();
                         editor.container.remove();
@@ -95,7 +113,7 @@ module.exports = {
                             snippet: "will: $1",
                             meta: "snippet",
                             command: "startAutocomplete",
-                            range: new Range(0, 4, 0, 6)
+                            range: new Range(0, 4, 0, 7)
                         }, {
                             caption: "here",
                             value: "-here",
@@ -110,12 +128,12 @@ module.exports = {
         editor.moveCursorTo(0, 6);
         sendKey("w");
         var popup = editor.completer.popup;
-        check(function () {
+        afterRenderCheck(popup, function () {
             assert.equal(popup.data.length, 1);
             editor.onCommandKey(null, 0, 13);
             assert.equal(popup.data.length, 2);
             assert.equal(editor.getValue(), "goodwill: ");
-            check(function () {
+            afterRenderCheck(popup, function () {
                 editor.onCommandKey(null, 0, 13);
                 assert.equal(editor.getValue(), "goodwill-here");
                 editor.destroy();
@@ -123,14 +141,79 @@ module.exports = {
                 done();
             });
         });
+    },
+    "test: correct completion replacement range when completion prefix has more than one letter": function (done) {
+        var editor = initEditor("<");
+
+        editor.completers = [
+            {
+                getCompletions: function (editor, session, pos, prefix, callback) {
+                    var completions = [
+                        {
+                            caption: "div",
+                            value: "div",
+                            range: new Range(0, 1, 0, 3)
+                        }, {
+                            caption: "dialog",
+                            value: "dialog",
+                            range: new Range(0, 1, 0, 3)
+                        }
+                    ];
+                    callback(null, completions);
+                }
+            }
+        ];
+
+        editor.moveCursorTo(0, 1);
+        sendKey("di");
+        var popup = editor.completer.popup;
+        check(function () {
+            assert.equal(popup.data.length, 2);
+            editor.onCommandKey(null, 0, 13);
+            check(function () {
+                assert.equal(editor.getValue(), "<dialog");
+                editor.destroy();
+                editor.container.remove();
+                done();
+            });
+        });
 
         function check(callback) {
-            popup = editor.completer.popup;
-            popup.renderer.on("afterRender", function wait() {
-                popup.renderer.off("afterRender", wait);
+            setTimeout(function wait() {
                 callback();
-            });
+            }, 10);
         }
+    },
+    "test: symbols after selection are not removed when replacement range is present": function (done) {
+        var editor = initEditor("{}");
+        editor.completers = [
+            {
+                getCompletions: function (editor, session, pos, prefix, callback) {
+                    var completions = [
+                        {
+                            caption: "apple",
+                            snippet: "apple: $1",
+                            meta: "snippet",
+                            range: new Range(0, 1, 0, 2)
+                        }, {
+                            caption: "pineapple",
+                            value: "pineapple",
+                            range: new Range(0, 1, 0, 2)
+                        }
+                    ];
+                    callback(null, completions);
+                }
+            }
+        ];
+        editor.moveCursorTo(0, 1);
+        sendKey("a");
+        var popup = editor.completer.popup;
+        afterRenderCheck(popup, function () {
+            assert.equal(popup.data.length, 2);
+            editor.onCommandKey(null, 0, 13);
+            assert.equal(editor.getValue(), "{apple: }");
+            done();
+            });
     },
     "test: different completers tooltips": function (done) {
         var editor = initEditor("");
@@ -183,6 +266,8 @@ module.exports = {
                 id: "secondCompleter"
             }
         ];
+        
+        
 
         sendKey("c");
         var popup = editor.completer.popup;
@@ -374,6 +459,325 @@ module.exports = {
         
         editor.destroy();
         editor.container.remove();
+    },
+    "test: selection should follow hovermarker if setSelectOnHover true": function(done) {
+        var editor = initEditor("hello world\n");
+        
+        editor.completers = [
+            {
+                getCompletions: function (editor, session, pos, prefix, callback) {
+                    var completions = [
+                        {
+                            caption: "option 1",
+                            value: "one"
+                        }, {
+                            caption: "option 2",
+                            value: "two"
+                        }, {
+                            caption: "option 3",
+                            value: "three"
+                        }
+                    ];
+                    callback(null, completions);
+                }
+            }
+        ];
+        
+        var completer = Autocomplete.for(editor);
+        completer.setSelectOnHover = true;
+        user.type("Ctrl-Space");
+        assert.equal(editor.completer.popup.isOpen, true);     
+        assert.equal(completer.popup.getRow(), 0);
+        
+        var text = completer.popup.renderer.content.childNodes[2];
+        var rect = text.getBoundingClientRect();
+
+        // We need two mouse events to trigger the updating of the hover marker.
+        text.dispatchEvent(new MouseEvent("move", {x: rect.left, y: rect.top}));
+        // Hover over the second row.
+        text.dispatchEvent(new MouseEvent("move", {x: rect.left + 1, y: rect.top + 20}));
+
+        // Selected row should follow mouse.
+        editor.completer.popup.renderer.$loop._flush();
+        assert.equal(completer.popup.getRow(), 2);
+
+        done();
+    }, 
+    "test: selection should not follow hovermarker if setSelectOnHover not set": function(done) {
+        var editor = initEditor("hello world\n");
+
+        editor.completers = [
+            {
+                getCompletions: function (editor, session, pos, prefix, callback) {
+                    var completions = [
+                        {
+                            caption: "option 1",
+                            value: "one"
+                        }, {
+                            caption: "option 2",
+                            value: "two"
+                        }, {
+                            caption: "option 3",
+                            value: "three"
+                        }
+                    ];
+                    callback(null, completions);
+                }
+            }
+        ];
+
+        var completer = Autocomplete.for(editor);
+ 
+        user.type("Ctrl-Space");
+        assert.equal(editor.completer.popup.isOpen, true);     
+        assert.equal(completer.popup.getRow(), 0);
+        
+        var text = completer.popup.renderer.content.childNodes[2];
+        var rect = text.getBoundingClientRect();
+
+        // We need two mouse events to trigger the updating of the hover marker.
+        text.dispatchEvent(new MouseEvent("move", {x: rect.left, y: rect.top}));
+        // Hover over the second row.
+        text.dispatchEvent(new MouseEvent("move", {x: rect.left + 1, y: rect.top + 20}));
+
+        // Selected row should not follow mouse.
+        editor.completer.popup.renderer.$loop._flush();
+        assert.equal(completer.popup.getRow(), 0);
+
+        done();
+    },
+    "test: should respect hideInlinePreview": function(done) {
+        var editor = initEditor("hello world\n");
+        
+        editor.completers = [
+            {
+                getCompletions: function (editor, session, pos, prefix, callback) {
+                    var completions = [
+                        {
+                            caption: "option 1",
+                            value: "one",
+                            score: 3
+                        }
+                    ];
+                    callback(null, completions);
+                },
+                hideInlinePreview: true
+            }, {
+                getCompletions: function (editor, session, pos, prefix, callback) {
+                    var completions = [
+                        {
+                            caption: "option 2",
+                            value: "two",
+                            score: 2
+                        }
+                    ];
+                    callback(null, completions);
+                },
+                hideInlinePreview: false
+            }, {
+                getCompletions: function (editor, session, pos, prefix, callback) {
+                    var completions = [
+                        {
+                            caption: "option 3",
+                            value: "three",
+                            score: 1
+                        }
+                    ];
+                    callback(null, completions);
+                }
+            }
+        ];
+        
+        var completer = Autocomplete.for(editor);
+        completer.inlineEnabled = true;
+
+        user.type("Ctrl-Space");
+        var inline = completer.inlineRenderer;
+
+        assert.equal(editor.completer.popup.isOpen, true);  
+        
+        // Row 0, should hide inline preview.
+        assert.equal(completer.popup.getRow(), 0);
+        assert.strictEqual(inline.isOpen(), false);
+
+        sendKey("Down");
+
+        // Row 1, should show inline preview.
+        assert.equal(completer.popup.getRow(), 1);
+        assert.strictEqual(inline.isOpen(), true);
+
+        sendKey("Down");
+
+        // Row 2, should show inline preview.
+        assert.equal(completer.popup.getRow(), 2);
+        assert.strictEqual(inline.isOpen(), true);
+
+
+        done();
+    },
+    "test: should maintain selection on fast completer item when slow completer results come in": function(done) {
+        var editor = initEditor("hello world\n");
+        
+        var slowCompleter = {
+            getCompletions: function (editor, session, pos, prefix, callback) {
+                var completions = [
+                    {
+                        caption: "slow option 1",
+                        value: "s1",
+                        score: 3
+                    }, {
+                        caption: "slow option 2",
+                        value: "s2",
+                        score: 0
+                    }
+                ];
+                setTimeout(() => {
+                    callback(null,  completions);
+                }, 200);
+            }
+        };
+
+        var fastCompleter = {
+            getCompletions: function (editor, session, pos, prefix, callback) {
+                var completions = [
+                    {
+                        caption: "fast option 1",
+                        value: "f1",
+                        score: 2
+                    }, {
+                        caption: "fast option 2",
+                        value: "f2",
+                        score: 1
+                    }
+                ];
+                callback(null, completions);
+            }
+        };
+
+        editor.completers = [fastCompleter, slowCompleter];
+        
+        var completer = Autocomplete.for(editor);
+        completer.stickySelectionDelay = 100;
+        user.type("Ctrl-Space");
+        assert.equal(completer.popup.isOpen, true);    
+        assert.equal(completer.popup.data.length, 2); 
+        assert.equal(completer.popup.getRow(), 0);
+
+        setTimeout(() => {
+            completer.popup.renderer.$loop._flush();
+            assert.equal(completer.popup.data.length, 4);
+            assert.equal(completer.popup.getRow(), 1);
+     
+            done();
+        }, 500);    
+    },
+    "test: should not maintain selection on fast completer item when slow completer results come in when stickySelectionDelay negative": function(done) {
+        var editor = initEditor("hello world\n");
+        
+        var slowCompleter = {
+            getCompletions: function (editor, session, pos, prefix, callback) {
+                var completions = [
+                    {
+                        caption: "slow option 1",
+                        value: "s1",
+                        score: 3
+                    }, {
+                        caption: "slow option 2",
+                        value: "s2",
+                        score: 0
+                    }
+                ];
+                setTimeout(() => {
+                    callback(null,  completions);
+                }, 200);
+            }
+        };
+
+        var fastCompleter = {
+            getCompletions: function (editor, session, pos, prefix, callback) {
+                var completions = [
+                    {
+                        caption: "fast option 1",
+                        value: "f1",
+                        score: 2
+                    }, {
+                        caption: "fast option 2",
+                        value: "f2",
+                        score: 1
+                    }
+                ];
+                callback(null, completions);
+            }
+        };
+
+        editor.completers = [fastCompleter, slowCompleter];
+        
+        var completer = Autocomplete.for(editor);
+        completer.stickySelectionDelay = -1;
+        user.type("Ctrl-Space");
+        assert.equal(completer.popup.isOpen, true);    
+        assert.equal(completer.popup.data.length, 2); 
+        assert.equal(completer.popup.getRow(), 0);
+
+        setTimeout(() => {
+            completer.popup.renderer.$loop._flush();
+            assert.equal(completer.popup.data.length, 4);
+            assert.equal(completer.popup.getRow(), 0);
+     
+            done();
+        }, 500);    
+    },
+    "test: should filter using caption if ignoreCaption false": function() {
+        var editor = initEditor("hello world\n");
+        
+        var completer = {
+            getCompletions: function (editor, session, pos, prefix, callback) {
+                var completions = [
+                    {
+                        caption: "caption",
+                        value: "value"
+                    }
+                ];
+                callback(null,  completions);
+            }
+        };
+
+        editor.completers = [completer];
+        
+        var completer = Autocomplete.for(editor);
+
+        // Should filter using the caption if set to false.
+        completer.ignoreCaption = false;
+        user.type("cap");
+        assert.equal(completer.popup.isOpen, true);   
+    },
+    "test: should filter using value if ignoreCaption true": function() {
+        var editor = initEditor("hello world\n");
+        
+        var completer = {
+            getCompletions: function (editor, session, pos, prefix, callback) {
+                var completions = [
+                    {
+                        caption: "caption",
+                        value: "value"
+                    }
+                ];
+                callback(null,  completions);
+            }
+        };
+
+        editor.completers = [completer];
+        
+        var completer = Autocomplete.for(editor);
+
+        // Should not filter using the caption if set to true.
+        completer.ignoreCaption = true;
+        user.type("cap");
+        assert.equal(completer.popup, undefined);   
+
+        // Should filter using the value instead.
+        user.type(" value");
+        assert.equal(completer.popup.isOpen, true);   
     }
 };