+++ src/keyboard/textinput_test.js
@@ -153,46 +153,6 @@ module.exports = {
         assert.equal(editor.getValue(), "y");
     },
     
-    "test: android spacebar moves cursor": function() {
-        setUserAgentForTests(true, false);
-        var value = "Juhu kinners!";
-        editor.setValue(value);
-        editor.blur();
-        editor.focus();
-        var lastCommand = "";
-        editor.commands.on("exec", function(e) {
-            lastCommand += e.command.name;
-        });
-        
-        textarea.selectionStart = textarea.selectionEnd;
-        document.dispatchEvent(new CustomEvent("selectionchange"));
-        assert.equal(lastCommand, "gotoright");
-        lastCommand = "";
-        
-        textarea.selectionStart = 
-        textarea.selectionEnd = textarea.selectionStart - 1;
-        document.dispatchEvent(new CustomEvent("selectionchange"));
-        assert.equal(lastCommand, "gotoleft");
-        lastCommand = "";
-        
-        assert.equal(editor.getSelectedText(), "");
-        textarea.selectionStart = 0;
-        textarea.selectionEnd = textarea.value.length;
-        textarea.dispatchEvent(new CustomEvent("select"));
-        assert.equal(editor.getSelectedText(), value);
-        
-        textarea.selectionEnd = textarea.selectionStart;
-        document.dispatchEvent(new CustomEvent("selectionchange"));
-        assert.equal(lastCommand, "gotoleft");
-        lastCommand = "";
-        
-        textarea.selectionStart = 
-        textarea.selectionEnd = textarea.selectionEnd + 2;
-        document.dispatchEvent(new CustomEvent("selectionchange"));
-        assert.equal(lastCommand, "gotorightgotoright");
-        lastCommand = "";
-    },
-    
     "test: composition with visible textarea": function() {
         var data = [
             // select ll
@@ -484,6 +444,40 @@ module.exports = {
             { _: "keydown", range: [0,0], value: "\n\n", key: { code: "ShiftLeft", key: "Shift", keyCode: 16}, modifier: "shift-"},
             { _: "keydown", range: [3,8], value: "kinners\n\n\n", key: { code: "ArrowUp", key: "ArrowUp", keyCode: 38}, modifier: "shift-"},
             { _: "keydown", range: [3,12], value: "juhu\nkinners\n\n", key: { code: "ArrowUp", key: "ArrowUp", keyCode: 38}, modifier: "shift-"}
+        ].forEach(function(data) {
+            sendEvent(data._, data);
+        });
+        // test overflow
+        editor.session.setValue("0123456789".repeat(80));
+        editor.execCommand("gotoright");
+        editor.execCommand("selectright");
+        assert.equal([textarea.value.length, textarea.selectionStart, textarea.selectionEnd].join(","), "402,1,2");
+        editor.execCommand("gotolineend");
+        assert.equal([textarea.value.length, textarea.selectionStart, textarea.selectionEnd].join(","), "3,0,0");
+        editor.execCommand("selectleft");
+        assert.equal([textarea.value.length, textarea.selectionStart, textarea.selectionEnd].join(","), "3,0,1");
+    },
+    
+    "test: selection synchronization with extra lines enabled": function() {
+        editor.textInput.setNumberOfExtraLines(1);
+        editor.session.setValue("line1\nline2\nline3\nline4\nline5\nline6\n");
+        [
+            { _: "keydown", range: [1,1], value: "line1\nline2\nline3\n\n", key: { code: "ArrowRight", key: "ArrowRight", keyCode: 39}},
+            { _: "keydown", range: [2,2], value: "line1\nline2\nline3\n\n", key: { code: "ArrowRight", key: "ArrowRight", keyCode: 39}},
+            { _: "keydown", range: [2,2], value: "line1\nline2\nline3\n\n", key: { code: "ShiftLeft", key: "Shift", keyCode: 16}, modifier: "shift-"},
+            { _: "keydown", range: [2,3], value: "line1\nline2\nline3\n\n", key: { code: "ArrowRight", key: "ArrowRight", keyCode: 39}, modifier: "shift-"},
+            { _: "keydown", range: [2,4], value: "line1\nline2\nline3\n\n", key: { code: "ArrowRight", key: "ArrowRight", keyCode: 39}, modifier: "shift-"},
+            { _: "keydown", range: [2,5], value: "line1\nline2\nline3\n\n", key: { code: "ArrowRight", key: "ArrowRight", keyCode: 39}, modifier: "shift-"},
+            { _: "keydown", range: [2,6], value: "line1\nline2\nline3\n\n", key: { code: "ArrowRight", key: "ArrowRight", keyCode: 39}, modifier: "shift-"},
+            { _: "keydown", range: [2,7], value: "line1\nline2\nline3\n\n", key: { code: "ArrowRight", key: "ArrowRight", keyCode: 39}, modifier: "shift-"},
+            { _: "keydown", range: [2,8], value: "line1\nline2\nline3\n\n", key: { code: "ArrowRight", key: "ArrowRight", keyCode: 39}, modifier: "shift-"},
+            { _: "keydown", range: [2,14], value: "line1\nline2\nline3\n\n", key: { code: "ArrowDown", key: "ArrowDown", keyCode: 40}, modifier: "shift-"},
+            { _: "keydown", range: [2,2], value: "line4\nline5\nline6\n\n", key: { code: "ArrowDown", key: "ArrowDown", keyCode: 40}},
+            { _: "keydown", range: [2,2], value: "line4\nline5\nline6\n\n", key: { code: "ShiftLeft", key: "Shift", keyCode: 16}, modifier: "shift-"},
+            { _: "keydown", range: [14,20], value: "line1\nline2\nline3\nline4\n\n", key: { code: "ArrowUp", key: "ArrowUp", keyCode: 38}, modifier: "shift-"},
+            { _: "keydown", range: [8,8], value: "line1\nline2\nline3\n\n", key: { code: "ArrowUp", key: "ArrowUp", keyCode: 38}},
+            { _: "keydown", range: [14,14], value: "line1\nline2\nline3\n\n", key: { code: "ArrowDown", key: "ArrowDown", keyCode: 40}},
+            { _: "keydown", range: [2,8], value: "line3\nline4\nline5\nline6\n\n", key: { code: "ArrowDown", key: "ArrowDown", keyCode: 40}, modifier: "shift-"}
         ].forEach(function(data) {
             sendEvent(data._, data);
         });