+++ src/test/util.js
@@ -1,0 +1,10 @@
+/**
+ * @returns {boolean} true if it's a Node.js environment, false otherwise
+ */
+function isNodeEnvironment() {
+    return typeof process !== "undefined";
+}
+
+module.exports = {
+    isNodeEnvironment
+};